feat(1769): Add processHooks endpoint for queue-service cooperation
