+++ lib/registerPlugins.js
@@ -60,7 +60,8 @@ async function registerResourcePlugins(server, config) {
         'isAdmin',
         'shutdown',
         'release',
-        'validator'
+        'validator',
+        'processHooks'
     ];
 
     if (hoek.reach(config, 'coverage.coveragePlugin')) {
