+++ test/lib/registerPlugins.test.js
@@ -39,7 +39,8 @@ describe('Register Unit Test Case', () => {
         '../plugins/isAdmin',
         '../plugins/shutdown',
         '../plugins/release',
-        '../plugins/validator'
+        '../plugins/validator',
+        '../plugins/processHooks'
     ];
     const authPlugins = ['@hapi/bell', '@hapi/cookie', '@hapi/crumb', 'hapi-auth-bearer-token', 'hapi-auth-jwt2'];
     const pluginLength = expectedPlugins.length + resourcePlugins.length + authPlugins.length; // for server.register of auth Plugins;
