+++ test/plugins/data/webhookConfigPush.json
@@ -1,0 +1,31 @@
+{
+    "action": "push",
+    "branch": "master",
+    "checkoutUrl": "git@github.com:baxterthehacker/public-repo.git",
+    "sha": "0d1a26e67d8f5eaf1f6ba5c57fc3c7d91ac0fd1c",
+    "type": "repo",
+    "username": "baxterthehacker2",
+    "commitAuthors": [
+        "baxterthehacker"
+    ],
+    "lastCommitMessage": "lastcommitmessage",
+    "hookId": "3c77bf80-9a2f-11e6-80d6-72f7fe03ea29",
+    "scmContext": "github:github.com",
+    "ref": "refs/heads/master",
+    "addedFiles": [
+        "README.md"
+    ],
+    "modifiedFiles": [
+        "README.md",
+        "package.json"
+    ],
+    "removedFiles": [
+        "screwdriver.yaml"
+    ],
+    "pluginOptions": {
+        "username": "sd-buildbot",
+        "ignoreCommitsBy": ["batman", "superman"],
+        "restrictPR": "fork",
+        "chainPR": false
+    }
+}
